WA-284 logger protocol
